Restructuring test environments strategy
