+++ requirements-dev.txt
@@ -1,3 +1,4 @@
+wheel~=0.34.2
 awscli>=1.18.0,<2.0.0
 black~=19.3b0
 pylint~=2.5.3
@@ -15,7 +16,6 @@ scikit-learn~=0.23.1
 cfn-lint~=0.33.1
 cfn-flip~=1.2.3
 twine~=3.1.1
-wheel~=0.34.2
 sphinx~=3.1.1
 sphinx_bootstrap_theme~=0.7.1
 moto~=1.3.14
